Move strategies to naive subpackage
